+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Goliac v1.3.1
+
+- support of apps in branch protection bypass and global ruleset
+
+## Goliac v1.3.0
+
+- introducing branch protection bypass (users/teams)
+
 ## Goliac v1.2.2
 
 - bugfix: support disabled ruleset